+++ c++/diag.hpp
@@ -16,8 +16,14 @@
 #include "params.hpp"
 #include "eigen.hpp"
 #include "time_mem.hpp"
-#include "numerics.hpp" // is_matrix_upper
 #include "debug.hpp" // nrglogdp
+
+
+#ifdef USE_UBLAS
+#include "numerics_ublas.hpp" // is_matrix_upper
+#else
+#include "numerics_Eigen.hpp" // is_matrix_upper
+#endif
 
 #define LAPACK_COMPLEX_STRUCTURE
 #include "lapack.h"
