+++ c++/step.hpp
@@ -3,7 +3,12 @@
 
 #include <algorithm>
 #include "params.hpp"
-#include "numerics.hpp"
+
+#ifdef USE_UBLAS
+#include "numerics_ublas.hpp"
+#else
+#include "numerics_Eigen.hpp"
+#endif
 
 #define FMT_HEADER_ONLY
 #include <fmt/format.h>
