+++ test/unit/misc/misc.cpp
@@ -148,45 +148,3 @@ TEST(misc, vector_of_keys){
 	const std::vector expected = {-10,12,42,99};
   ASSERT_EQ(b, expected);
 }
-
-
-
-
-
-
-TEST(misc, ublas_to_eigen){
-  auto ublas_matrix = read_matrix("txt/matrix.txt");
-  auto eigen_matrix = ublas_to_eigen(ublas_matrix);
-  compare(eigen_matrix, ublas_matrix);
-
-  ublas::vector<int> ublas_vector(5);
-  for(int i = 0; i < 5; i++) ublas_vector(i) = i;
-  auto eigen_vector = ublas_to_eigen(ublas_vector);
-  compare(eigen_vector, ublas_vector);
-}
-
-TEST(misc, eigen_to_ublas){
-  Eigen::Vector4i eigen_vector(3,5,8,14);
-  auto ublas_vector = eigen_to_ublas_vector(eigen_vector);
-  compare(eigen_vector, ublas_vector);
-  {
-    Eigen::Matrix3i eigen_matrix;
-    for (int i = 0; i < eigen_matrix.size(); i++) eigen_matrix(i) = i;
-    auto ublas_matrix = eigen_to_ublas_matrix(eigen_matrix);
-    compare(eigen_matrix, ublas_matrix);
-  }
-  
-  {
-    Eigen::Matrix<int, 3, 3, Eigen::RowMajor> eigen_matrix;
-    for (int i = 0; i < eigen_matrix.size(); i++) eigen_matrix(i) = i;
-    auto ublas_matrix = eigen_to_ublas_matrix(eigen_matrix);
-    compare(eigen_matrix, ublas_matrix);
-  }
-}
-
-
-
-int main(int argc, char **argv) {
-   ::testing::InitGoogleTest(&argc, argv);
-   return RUN_ALL_TESTS();
-}
