+++ test/unit/numerics/txt/matrix.txt
@@ -1,0 +1,3 @@
+31  41  53  46
+12  5   1   41
+5   2   4   7
